multiple changes  * centralized config  * better sharing of functions
