+++ src/model/BoxPlotColumn.ts
@@ -13,7 +13,6 @@ import NumberColumn, {
   INumberFilter,
   noNumberFilter
 } from './NumberColumn';
-import {isSortedByMe} from '../utils';
 
 export const SORT_METHOD = {
   min: 'min',
@@ -178,7 +177,7 @@ export default class BoxPlotColumn extends ValueColumn<IBoxPlotData> implements 
     }
     this.fire([Column.EVENT_SORTMETHOD_CHANGED], this.sort, this.sort = sort);
     // sort by me if not already sorted by me
-    if (!isSortedByMe(this)) {
+    if (!this.isSortedByMe().asc) {
       this.sortByMe();
     }
   }
